TDR-2905 Add details to text area for description.
This ticket asks for the GDS details component to be added to the text
area for description. As far as I know, there isn't anything suitable in
DisplayProperties to deal with this so I've added two new rows for the
description field,  DetailsSummary and DetailsText and used them to add
the text to the form.