Validate the metadata on continue.
I've added a form into the metadata start page and a submit button
instead of a link. This posts to a validate method which fetches the
file statuses and counts the number of incomplete ones.
This is then sent to the front end to show the error message.
I've also copied the css from the prototype PR and updated some of the
class names.
Tests to follow