Return error if backend trigger is non-200 response
The transfer fails if the backend checks are not triggered. This can occur if a user PLAY_SESSION timeouts after an hour. This throws a 401 error on the request to trigger the backend check

The non-200 errors from the request were no being caught and users were directed to file checks results page, making it appear the the transfer is proceeding successfully, but the file checks page just hangs indefinitely.

This causes the users' confusion

Now if a non-200 response is received this causes an error to be returned and the user is redirected to the error page making it clearer that there has been a problem